requirements.txt
fastapi==0.75.1
uvicorn==0.17.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
pymongo==4.0
pymongo[srv]
requests
mitreattack-python